requirements.txt
absl-py==0.13.0
addict==2.4.0
aio-pika==6.8.2
aiofiles==0.8.0
aiohttp==3.7.4
aiormq==3.3.1
anyio==3.6.1
appdirs==1.4.4
APScheduler==3.6.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
attr==0.3.1
attrs==21.2.0
AutoTS==0.4.0
azure-core==1.26.1
azure-storage-blob==12.14.1
Babel==2.10.3
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bidict==0.22.0
bleach==5.0.1
blis==0.7.9
boto==2.49.0
boto3==1.16.63
botocore==1.19.63
boxing==0.1.4
CacheControl==0.12.11
cachetools==4.2.2
catalogue==2.0.8
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.0.3
cloudpickle==1.6.0
cmake==3.22.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.0.4
confection==0.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cycler==0.11.0
cymem==2.0.7
dask==2021.11.2
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
discopy==0.5.0
discord.py==2.1.0
Django==3.2.16
django-annoying==0.10.6
django-cors-headers==3.6.0
django-debug-toolbar==3.2.1
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rest-swagger==2.2.0
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
dnspython==1.16.0
docopt==0.6.2
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-yasg==1.20.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
expiringdict==1.2.2
fast_ctc_decode==0.3.0
fastjsonschema==2.16.1
fbmessenger==6.0.0
filelock==3.6.0
fire==0.4.0
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
flatbuffers==2.0
fonttools==4.28.5
fortls==2.12.0
fsspec==2022.5.0
future==0.18.2
gast==0.4.0
google-api-core==2.10.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.3.2
google-cloud-logging==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
graphviz==0.20.1
greenlet==1.1.2
grpc-google-iam-v1==0.12.3
grpcio==1.51.1
grpcio-status==1.48.2
h11==0.12.0
h5py==3.6.0
htmlmin==0.1.12
httptools==0.4.0
huggingface-hub==0.4.0
humanfriendly==10.0
idna==3.3
imagecodecs==2022.2.22
imageio==2.15.0
importlib-metadata==4.10.1
inflection==0.5.1
iniconfig==1.1.1
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
isodate==0.6.1
itsdangerous==2.0.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.0.1
json5==0.9.8
jsonpickle==2.0.0
jsonschema==3.2.0
jstyleson==0.0.2
jupyter-server==1.18.1
jupyter_client==7.4.5
jupyter_core==5.0.0
jupyterlab==3.4.4
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
kafka-python==2.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
label-studio-converter==0.0.48rc0
label-studio-tools==0.0.1
lambeq==0.2.8
langcodes==3.3.0
lark-parser==0.12.0
launchdarkly-server-sdk==7.5.0
libclang==13.0.0
lmdb==1.3.0
locket==1.0.0
lockfile==0.12.2
lxml==4.7.1
Markdown==3.3.6
MarkupSafe==2.0.1
mathparse==0.1.2
matplotlib==3.3.4
matplotlib-inline==0.1.6
mattermostwrapper==2.2
mistune==0.8.4
mpmath==1.2.1
msgpack==1.0.3
msrest==0.7.1
multidict==5.2.0
murmurhash==1.0.9
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.6
networkx==2.6.3
nibabel==3.2.2
nltk==3.6.7
notebook==6.4.12
notebook-shim==0.1.0
ntlm-auth==1.5.0
numpy==1.19.5
oauthlib==3.2.0
openai==0.18.1
openapi-codec==1.3.2
opencv-python==4.5.5.64
openpyxl==3.0.9
openvino==2022.1.0
openvino-dev==2022.1.0
openvino-telemetry==2022.1.1
opt-einsum==3.3.0
ordered-set==4.0.2
outcome==1.1.0
packaging==20.9
pamqp==2.3.0
pandas==1.3.5
pandas-datareader==0.10.0
pandas-stubs==1.2.0.58
pandocfilters==1.5.0
parasail==1.2.4
parso==0.8.3
partd==1.2.0
pathy==0.10.1
patsy==0.5.2
pbr==5.11.1
pickleshare==0.7.5
Pillow==9.3.0
Pint==0.18
platformdirs==2.5.4
plotly==5.6.0
pluggy==1.0.0
ply==3.11
preshed==3.0.8
progress==1.6
prometheus-client==0.14.1
prompt-toolkit==2.0.10
proto-plus==1.22.1
protobuf==3.20.3
psutil==5.9.4
psycopg2-binary==2.9.1
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post2
pycparser==2.21
pydantic==1.8.2
pydicom==2.3.0
pydot==1.4.2
Pygments==2.13.0
PyJWT==2.1.0
pykwalify==1.8.0
pymongo==3.10.1
PyMsgBox==1.0.9
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyperclip==1.8.2
pyreadline3==3.4.1
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.18.1
pyTelegramBotAPI==3.8.3
pytest==6.2.5
python-crfsuite==0.9.8
python-dateutil==2.8.1
python-dotenv==0.21.0
python-engineio==4.3.2
python-json-logger==2.0.4
python-socketio==5.6.0
python-telegram-bot==13.10
pytket==1.10.0
pytweening==1.0.4
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
pywin32==305
pywinpty==2.0.6
PyYAML==6.0
pyzmq==24.0.1
qiskit==0.39.5
qiskit-aer==0.11.2
qiskit-ibmq-provider==0.19.2
qiskit-terra==0.22.4
questionary==1.10.0
qwasm==1.0.1
radian==0.6.3
randomname==0.1.5
rasa==3.1.0
rasa-sdk==3.1.1
rawpy==0.17.1
rchitect==0.3.36
redis==3.5.3
regex==2021.8.28
requests==2.26.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
retworkx==0.12.1
rocketchat-API==1.16.0
rq==1.10.1
rsa==4.8
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
rules==2.2
rustworkx==0.12.1
s3transfer==0.3.7
sacremoses==0.0.49
sanic==21.12.1
Sanic-Cors==2.0.1
sanic-jwt==1.7.0
sanic-routing==0.7.2
scikit-image==0.19.2
scikit-learn==0.24.2
scipy==1.10.0
seaborn==0.11.2
selenium==4.1.0
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.3.1
Shapely==1.8.1.post1
simplejson==3.18.0
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.4
smart-open==6.3.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
spacy==3.5.0
spacy-legacy==3.0.11
spacy-loggers==1.0.4
SQLAlchemy==1.4.36
sqlparse==0.4.2
srsly==2.4.5
stack-data==0.6.1
statsmodels==0.13.2
stevedore==4.1.1
sympy==1.11.1
tabulate==0.8.9
tarsafe==0.0.3
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-addons==0.15.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.23.1
tensornetwork==0.4.6
termcolor==1.1.0
terminado==0.15.0
terminaltables==3.1.10
texttable==1.6.4
thinc==8.1.7
threadpoolctl==3.1.0
tifffile==2022.3.25
tinycss2==1.1.1
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.2
torch==1.13.1
torchaudio==0.13.1
torchtext==0.14.1
tornado==6.2
tqdm==4.62.3
traitlets==5.5.0
transformers==4.16.2
trio==0.19.0
trio-websocket==0.9.2
tweedledum==1.1.1
twilio==6.50.1
typeguard==2.13.3
typer==0.7.0
types-pkg-resources==0.1.3
typing-extensions==3.10.0.2
typing-utils==0.1.0
tzdata==2021.5
tzlocal==4.1
ua-parser==0.16.1
ujson==4.3.0
uritemplate==4.1.1
urllib3==1.26.7
user-agents==2.2.0
wasabi==1.1.1
wcwidth==0.2.5
webencodings==0.5.1
webexteamssdk==1.6
websocket-client==1.3.3
websockets==10.3
Werkzeug==2.0.2
wrapt==1.13.3
wsproto==1.0.0
xmljson==0.2.0
yarl==1.7.2
zipp==3.7.0
